agent_client/requirements.txt
websocket-client
mss
pynput
pillow
python-dotenv
